Committing a whole bunch of minor changes to get rid of compiler warnings that were the result of compiling with -Wall -Wshadow.  Still have the following warnings that I wasn't easily able to fix.  All of them are either unused variables or functions:
enum_scanner.cc:1235:17: warning: 'void yyunput(int, char*)' defined but not used [-Wunused-function]
config.tab.yy:548:8: warning: 'Number do_integer_op(char, const Number&, const Number&)' defined but not used [-Wunused-function]
config_scanner.cc:1536:17: warning: 'void yyunput(int, char*)' defined but not used [-Wunused-function]
builtin.cc:452:8: warning: 'double my_F_to_C(double)' defined but not used [-Wunused-function]
builtin.cc:464:8: warning: 'double my_C_to_F(double)' defined but not used [-Wunused-function]
grib_utils.cc:269:6: warning: 'void gds_to_rotated_latlon(const Section2_Header&, RotatedLatLonData&)' defined but not used [-Wunused-function]
color_scanner.cc:1269:17: warning: 'void yyunput(int, char*)' defined but not used [-Wunused-function]
gsi_util.h:86:13: warning: 'not_has_FCST_VALID_BEG' defined but not used [-Wunused-variable]
gsi_util.h:87:13: warning: 'not_has_FCST_VALID_END' defined but not used [-Wunused-variable]
gsi_util.h:88:13: warning: 'not_has_OBS_VALID_BEG' defined but not used [-Wunused-variable]
gsi_util.h:89:13: warning: 'not_has_OBS_VALID_END' defined but not used [-Wunused-variable]
gsi_util.h:90:13: warning: 'not_has_FCST_VAR' defined but not used [-Wunused-variable]
gsi_util.h:91:13: warning: 'not_has_OBS_VAR' defined but not used [-Wunused-variable]
gsi_util.h:92:13: warning: 'not_has_OBTYPE' defined but not used [-Wunused-variable]





Former-commit-id: 7d47b7279b9d59ef5225ba2724ec636f0e252842

+++ met/src/libcode/vx_tc_util/prob_info_array.cc
@@ -159,7 +159,7 @@ const ProbInfoBase * ProbInfoArray::operator[](int n) const {
 const ProbRIRWInfo & ProbInfoArray::prob_rirw(int n) const {
 
    // Check range
-   if((n < 0) || (n >= ProbRIRW.size())) {
+   if((n < 0) || (n >= (int) ProbRIRW.size())) {
       mlog << Error
            << "\nProbRIRWInfo & ProbInfoArray::prob_rirw(int) -> "
            << "range check error for index value " << n << "\n\n";
